requirements.txt
Jinja2==2.10
requests==2.20.0
colorama==0.3.9
six==1.11.0
numpy==1.15.0
xlrd==1.1.0
urllib3==1.23
xlwt==1.3.0
jieba==0.39
PyYAML==5.1
